+++ src/app/(main)/main/_components/hooks/useLinkModalOnce.ts
@@ -1,0 +1,14 @@
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+
+export const useLinkModalOnce = () => {
+  const router = useRouter();
+
+  useEffect(() => {
+    const shown = localStorage.getItem("link-modal-shown");
+    if (!shown) {
+      router.replace("/link");
+      localStorage.setItem("link-modal-shown", "true");
+    }
+  }, []);
+};
