+++ src/app/(mypage)/connect-complete/_components/CompleteContent.tsx
@@ -1,10 +1,14 @@
 "use client";
 
+import { Button } from "@/components/common/button";
 import { useSelectedSocialStore } from "../../connect/_components/store/social-store";
 import SnsProfile, { SnsType } from "../../mypage/_components/SnsProfile";
+import { ArrowIcon } from "@/components/icons/service";
+import { useRouter } from "next/navigation";
 
 export const CompleteContent = () => {
   const { selected } = useSelectedSocialStore();
+  const router = useRouter();
 
   return (
     <section className="pt-[194px] flex flex-col gap-5 items-center justify-center">
@@ -18,6 +22,17 @@ export const CompleteContent = () => {
       <div className="flex flex-col items-center gap-2">
         <p className="text-h3 text-grayscale-900">계정이 연결되었어요!</p>
         <p className="text-b2R text-grayscale-700">재플리와 함께 계정을 성장시켜보세요.</p>
+      </div>
+      <div className="fixed bottom-[60px] left-1/2 -translate-x-1/2 w-full max-w-[440px] px-5 z-50 flex flex-col gap-3">
+        <Button
+          className="w-full"
+          rightIcon={<ArrowIcon type="right" />}
+          onClick={() => router.push("/connect")}>
+          다른 계정 연결하기
+        </Button>
+        <Button className="w-full" variant="subAction" onClick={() => router.push("/mypage")}>
+          닫기
+        </Button>
       </div>
     </section>
   );
