+++ src/app/(mypage)/connect-complete/_components/ErrorContent.tsx
@@ -1,8 +1,14 @@
+"use client";
+
+import { Button } from "@/components/common/button";
 import { useSelectedSocialStore } from "../../connect/_components/store/social-store";
 import SnsProfile, { SnsType } from "../../mypage/_components/SnsProfile";
+import { ArrowIcon } from "@/components/icons/service";
+import { useRouter } from "next/navigation";
 
 export const ErrorContent = () => {
   const { selected } = useSelectedSocialStore();
+  const router = useRouter();
 
   return (
     <section className="pt-[194px] flex flex-col gap-5 items-center justify-center">
@@ -16,6 +22,17 @@ export const ErrorContent = () => {
       <div className="flex flex-col items-center gap-2">
         <p className="text-h3 text-grayscale-900">계정이 연결에 실패했어요.</p>
         <p className="text-b2R text-grayscale-700">다시 시도해주세요.</p>
+      </div>
+      <div className="fixed bottom-[60px] left-1/2 -translate-x-1/2 w-full max-w-[440px] px-5 z-50 flex flex-col gap-3">
+        <Button
+          className="w-full"
+          rightIcon={<ArrowIcon type="right" />}
+          onClick={() => router.push("/connect")}>
+          다시 시도하기
+        </Button>
+        <Button className="w-full" variant="subAction" onClick={() => router.push("/mypage")}>
+          닫기
+        </Button>
       </div>
     </section>
   );
